--- a/src/libraries/System.Formats.Asn1/src/System/Formats/Asn1/SetOfValueComparer.cs
+++ b/src/libraries/System.Formats.Asn1/src/System/Formats/Asn1/SetOfValueComparer.cs
@@ -17,6 +17,18 @@ namespace System.Formats.Asn1
             int min = Math.Min(x.Length, y.Length);
             int diff;
 
+#if NET7_0_OR_GREATER
+            int diffIndex = x.Slice(0, min).CommonPrefixLength(y.Slice(0, min));
+
+            if (diffIndex == min)
+            {
+                diff = 0;
+            }
+            else
+            {
+                return (int)x[diffIndex] - y[diffIndex];
+            }
+#else
             for (int i = 0; i < min; i++)
             {
                 int xVal = x[i];
@@ -28,6 +40,7 @@ namespace System.Formats.Asn1
                     return diff;
                 }
             }
+#endif
 
             // The sorting rules (T-REC-X.690-201508 sec 11.6) say that the shorter one
             // counts as if it are padded with as many 0x00s on the right as required for